Move MM client to seperate package
