+++ mmclient/http.go
@@ -1,35 +1,19 @@
-package proxy
+package mmclient
 
 import (
 	pluginapi "github.com/mattermost/mattermost-plugin-api"
 	"github.com/mattermost/mattermost-server/v5/model"
 
+	"github.com/mattermost/mattermost-plugin-apps/server/config"
 	"github.com/mattermost/mattermost-plugin-apps/utils"
 )
-
-type MMClient interface {
-	GetUserByUsername(userName string) (*model.User, error)
-	CreateUserAccessToken(userID, description string) (*model.UserAccessToken, error)
-	RevokeUserAccessToken(tokenID string) error
-
-	CreateOAuthApp(app *model.OAuthApp) error
-	GetOAuthApp(appID string) (*model.OAuthApp, error)
-	DeleteOAuthApp(appID string) error
-
-	GetBot(botUserID string) (*model.Bot, error)
-	CreateBot(bot *model.Bot) error
-	EnableBot(botUserID string) (*model.Bot, error)
-	DisableBot(botUserID string) (*model.Bot, error)
-}
 
 type httpClient struct {
 	mm *model.Client4
 }
 
-func (p *Proxy) GetMMHTTPClient(sessionID, actingUserID string) (MMClient, error) {
-	conf := p.conf.GetConfig()
-
-	client, err := utils.ClientFromSession(p.mm, conf.MattermostSiteURL, sessionID, actingUserID)
+func NewHTTPClient(mm *pluginapi.Client, conf config.Config, sessionID, actingUserID string) (Client, error) {
+	client, err := utils.ClientFromSession(mm, conf.MattermostSiteURL, sessionID, actingUserID)
 	if err != nil {
 		return nil, err
 	}
@@ -135,58 +119,4 @@ func (h *httpClient) DisableBot(botUserID string) (*model.Bot, error) {
 	}
 
 	return bot, nil
-}
-
-type rpcClient struct {
-	mm *pluginapi.Client
-}
-
-func (p *Proxy) GetMMRPCClient() MMClient {
-	return &rpcClient{p.mm}
-}
-
-// User section
-
-func (r *rpcClient) GetUserByUsername(userName string) (*model.User, error) {
-	return r.mm.User.GetByUsername(userName)
-}
-
-func (r *rpcClient) CreateUserAccessToken(userID, description string) (*model.UserAccessToken, error) {
-	return r.mm.User.CreateAccessToken(userID, description)
-}
-
-func (r *rpcClient) RevokeUserAccessToken(tokenID string) error {
-	return r.mm.User.RevokeAccessToken(tokenID)
-}
-
-// OAuth section
-
-func (r *rpcClient) CreateOAuthApp(app *model.OAuthApp) error {
-	return r.mm.OAuth.Create(app)
-}
-
-func (r *rpcClient) GetOAuthApp(appID string) (*model.OAuthApp, error) {
-	return r.mm.OAuth.Get(appID)
-}
-
-func (r *rpcClient) DeleteOAuthApp(appID string) error {
-	return r.mm.OAuth.Delete(appID)
-}
-
-// Bpt section
-
-func (r *rpcClient) CreateBot(bot *model.Bot) error {
-	return r.mm.Bot.Create(bot)
-}
-
-func (r *rpcClient) GetBot(botUserID string) (*model.Bot, error) {
-	return r.mm.Bot.Get(botUserID, false)
-}
-
-func (r *rpcClient) EnableBot(botUserID string) (*model.Bot, error) {
-	return r.mm.Bot.UpdateActive(botUserID, true)
-}
-
-func (r *rpcClient) DisableBot(botUserID string) (*model.Bot, error) {
-	return r.mm.Bot.UpdateActive(botUserID, false)
 }
