+++ server/proxy/service.go
@@ -12,6 +12,7 @@ import (
 	"github.com/mattermost/mattermost-plugin-api/cluster"
 
 	"github.com/mattermost/mattermost-plugin-apps/apps"
+	"github.com/mattermost/mattermost-plugin-apps/mmclient"
 	"github.com/mattermost/mattermost-plugin-apps/server/config"
 	"github.com/mattermost/mattermost-plugin-apps/server/httpout"
 	"github.com/mattermost/mattermost-plugin-apps/server/store"
@@ -44,21 +45,18 @@ type Service interface {
 
 	AddLocalManifest(actingUserID string, m *apps.Manifest) (md.MD, error)
 	AppIsEnabled(app *apps.App) bool
-	EnableApp(client MMClient, sessionID string, cc *apps.Context, appID apps.AppID) (md.MD, error)
-	DisableApp(client MMClient, sessionID string, cc *apps.Context, appID apps.AppID) (md.MD, error)
+	EnableApp(client mmclient.Client, sessionID string, cc *apps.Context, appID apps.AppID) (md.MD, error)
+	DisableApp(client mmclient.Client, sessionID string, cc *apps.Context, appID apps.AppID) (md.MD, error)
 	GetInstalledApp(appID apps.AppID) (*apps.App, error)
 	GetInstalledApps() []*apps.App
 	GetListedApps(filter string, includePluginApps bool) []*apps.ListedApp
 	GetManifest(appID apps.AppID) (*apps.Manifest, error)
 	GetManifestFromS3(appID apps.AppID, version apps.AppVersion) (*apps.Manifest, error)
-	InstallApp(client MMClient, sessionID string, cc *apps.Context, trusted bool, secret, pluginID string) (*apps.App, md.MD, error)
+	InstallApp(client mmclient.Client, sessionID string, cc *apps.Context, trusted bool, secret, pluginID string) (*apps.App, md.MD, error)
 	SynchronizeInstalledApps() error
-	UninstallApp(client MMClient, sessionID string, cc *apps.Context, appID apps.AppID) (md.MD, error)
+	UninstallApp(client mmclient.Client, sessionID string, cc *apps.Context, appID apps.AppID) (md.MD, error)
 
 	AddBuiltinUpstream(apps.AppID, upstream.Upstream)
-
-	GetMMHTTPClient(sessionID, actingUserID string) (MMClient, error)
-	GetMMRPCClient() MMClient
 }
 
 var _ Service = (*Proxy)(nil)
